requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
